Migrate Parquet dictionary test cases to presto-parquet
