+++ presto-hive/src/test/java/io/prestosql/plugin/hive/parquet/predicate/TestParquetPredicateUtils.java
@@ -15,7 +15,6 @@ package io.prestosql.plugin.hive.parquet.predicate;
 
 import com.google.common.collect.ImmutableList;
 import com.google.common.collect.ImmutableMap;
-import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Iterables;
 import io.prestosql.parquet.RichColumnDescriptor;
 import io.prestosql.plugin.hive.HiveColumnHandle;
@@ -27,7 +26,6 @@ import io.prestosql.spi.type.MapType;
 import io.prestosql.spi.type.RowType;
 import io.prestosql.spi.type.StandardTypes;
 import org.apache.parquet.column.ColumnDescriptor;
-import org.apache.parquet.column.Encoding;
 import org.apache.parquet.schema.GroupType;
 import org.apache.parquet.schema.MessageType;
 import org.apache.parquet.schema.PrimitiveType;
@@ -36,11 +34,8 @@ import org.testng.annotations.Test;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
 
-import static com.google.common.collect.Sets.union;
 import static io.prestosql.parquet.ParquetTypeUtils.getDescriptors;
-import static io.prestosql.parquet.predicate.PredicateUtils.isOnlyDictionaryEncodingPages;
 import static io.prestosql.plugin.hive.HiveColumnHandle.ColumnType.REGULAR;
 import static io.prestosql.plugin.hive.parquet.ParquetPageSourceFactory.getParquetTupleDomain;
 import static io.prestosql.spi.block.MethodHandleUtil.methodHandle;
@@ -48,44 +43,16 @@ import static io.prestosql.spi.predicate.TupleDomain.withColumnDomains;
 import static io.prestosql.spi.type.BigintType.BIGINT;
 import static io.prestosql.spi.type.IntegerType.INTEGER;
 import static io.prestosql.spi.type.TypeSignature.parseTypeSignature;
-import static org.apache.parquet.column.Encoding.BIT_PACKED;
-import static org.apache.parquet.column.Encoding.PLAIN;
-import static org.apache.parquet.column.Encoding.PLAIN_DICTIONARY;
-import static org.apache.parquet.column.Encoding.RLE;
 import static org.apache.parquet.schema.PrimitiveType.PrimitiveTypeName.INT32;
 import static org.apache.parquet.schema.PrimitiveType.PrimitiveTypeName.INT64;
 import static org.apache.parquet.schema.Type.Repetition.OPTIONAL;
 import static org.apache.parquet.schema.Type.Repetition.REPEATED;
 import static org.apache.parquet.schema.Type.Repetition.REQUIRED;
 import static org.testng.Assert.assertEquals;
-import static org.testng.Assert.assertFalse;
 import static org.testng.Assert.assertTrue;
 
 public class TestParquetPredicateUtils
 {
-    @Test
-    @SuppressWarnings("deprecation")
-    public void testDictionaryEncodingCasesV1()
-    {
-        Set<Encoding> required = ImmutableSet.of(BIT_PACKED);
-        Set<Encoding> optional = ImmutableSet.of(BIT_PACKED, RLE);
-        Set<Encoding> repeated = ImmutableSet.of(RLE);
-
-        Set<Encoding> notDictionary = ImmutableSet.of(PLAIN);
-        Set<Encoding> mixedDictionary = ImmutableSet.of(PLAIN_DICTIONARY, PLAIN);
-        Set<Encoding> dictionary = ImmutableSet.of(PLAIN_DICTIONARY);
-
-        assertFalse(isOnlyDictionaryEncodingPages(union(required, notDictionary)), "required notDictionary");
-        assertFalse(isOnlyDictionaryEncodingPages(union(optional, notDictionary)), "optional notDictionary");
-        assertFalse(isOnlyDictionaryEncodingPages(union(repeated, notDictionary)), "repeated notDictionary");
-        assertFalse(isOnlyDictionaryEncodingPages(union(required, mixedDictionary)), "required mixedDictionary");
-        assertFalse(isOnlyDictionaryEncodingPages(union(optional, mixedDictionary)), "optional mixedDictionary");
-        assertFalse(isOnlyDictionaryEncodingPages(union(repeated, mixedDictionary)), "repeated mixedDictionary");
-        assertTrue(isOnlyDictionaryEncodingPages(union(required, dictionary)), "required dictionary");
-        assertTrue(isOnlyDictionaryEncodingPages(union(optional, dictionary)), "optional dictionary");
-        assertTrue(isOnlyDictionaryEncodingPages(union(repeated, dictionary)), "repeated dictionary");
-    }
-
     @Test
     public void testParquetTupleDomainPrimitiveArray()
     {
